Split goal-funding.spec.ts into goal-funding-tests/ directory
The single spec file had grown to 449 lines, over the 360-line test-file
split guideline. Break it into one file per exported function with shared
fixtures, per the {module}-tests/ convention:

- fixtures.ts — makeTransaction, makeGoal, REF_DATE
- compute-monthly-deposit-rate.spec.ts
- compute-goal-eta.spec.ts
- compute-zipf-shares.spec.ts
- compute-goal-eta-from-share.spec.ts

Pure refactor — no test logic changes; all 31 tests still pass.
